Refactor Tokenizer to their own namespace
