+++ tests/tokenizer/BigramTokenizerTest.php
@@ -1,6 +1,9 @@
 <?php
 
-use TeamTNT\TNTSearch\Support\BigramTokenizer;
+namespace tokenizer;
+
+use PHPUnit;
+use TeamTNT\TNTSearch\Tokenizer\BigramTokenizer;
 
 class BigramTokenizerTest extends PHPUnit\Framework\TestCase
 {
@@ -9,7 +12,7 @@ class BigramTokenizerTest extends PHPUnit\Framework\TestCase
         $tokenizer = new BigramTokenizer;
 
         $text = "Quick Foxes";
-        $res  = $tokenizer->tokenize($text);
+        $res = $tokenizer->tokenize($text);
 
         $this->assertEquals(["qu", "ui", "ic", "ck", "fo", "ox", "xe", "es"], $res);
     }
