+++ tests/tokenizer/NGramTokenizerTest.php
@@ -1,6 +1,9 @@
 <?php
 
-use TeamTNT\TNTSearch\Support\NGramTokenizer;
+namespace tokenizer;
+
+use PHPUnit;
+use TeamTNT\TNTSearch\Tokenizer\NGramTokenizer;
 
 class NGramTokenizerTest extends PHPUnit\Framework\TestCase
 {
@@ -9,7 +12,7 @@ class NGramTokenizerTest extends PHPUnit\Framework\TestCase
         $tokenizer = new NGramTokenizer(3, 3);
 
         $text = "Quick Foxes";
-        $res  = $tokenizer->tokenize($text);
+        $res = $tokenizer->tokenize($text);
 
         $this->assertEquals(["qui", "uic", "ick", "fox", "oxe", "xes"], $res);
     }
@@ -19,7 +22,7 @@ class NGramTokenizerTest extends PHPUnit\Framework\TestCase
         $tokenizer = new NGramTokenizer(1, 2);
 
         $text = "Quick Fox";
-        $res  = $tokenizer->tokenize($text);
+        $res = $tokenizer->tokenize($text);
 
         $this->assertEquals(["q", "u", "i", "c", "k", "qu", "ui", "ic", "ck", "f", "o", "x", "fo", "ox"], $res);
     }
@@ -29,7 +32,7 @@ class NGramTokenizerTest extends PHPUnit\Framework\TestCase
         $tokenizer = new NGramTokenizer(4, 4);
 
         $text = "Quick Foxes";
-        $res  = $tokenizer->tokenize($text);
+        $res = $tokenizer->tokenize($text);
 
         $this->assertEquals(["quic", "uick", "foxe", "oxes"], $res);
     }
